Logout implemented
