+++ client/reducers/index.js
@@ -67,7 +67,10 @@ const loginUser = (state = {}, action) => {
       return Object.assign({}, state, {
         userID: action.data.id,
         userName: action.data.name,
+        status: true,
       });
+    case 'LOGOUT_USER':
+      return {};
     default:
       return state;
   }
