+++ server/routes.js
@@ -2,29 +2,10 @@ const User = require('./db/controllers/usersController.js');
 const Meal = require('./db/controllers/mealsController.js');
 const Order = require('./db/controllers/ordersController.js');
 const Controller = require('./db/controllers/controller.js');
-const passport = require('passport');
+const Auth = require('./db/controllers/authController.js');
+
 
 module.exports = (app) => {
-  const attemptSignIn = (req, res, user) => {
-    req.login(user, err => {
-      if (err) {
-        res.status(401).send(err);
-      } else {
-        res.json(user);
-      }
-    });
-  };
-
-  const signIn = (req, res, next) => {
-    passport.authenticate('local', (err, user, info) => {
-      if (err || !user) {
-        res.status(401).send(info);
-      } else {
-        attemptSignIn(req, res, user);
-      }
-    })(req, res, next);
-  };
-
 
   app.post('/api/users', User.createUser);
   app.get('/api/users', User.getAllUsers);
@@ -40,12 +21,8 @@ module.exports = (app) => {
 
   app.get('/api/meal/:id', Controller.getMealView);
 
-  app.post('/api/auth/sign-in', signIn);
-
-  app.get('/logout', (req, res) => {
-    req.logout();
-    res.redirect('/');
-  });
+  app.post('/api/auth/sign-in', Auth.signIn);
+  app.get('/signout', Auth.logOut);
 
   const isLoggedIn = (req, res, next) => {
     if (req.isAuthenticated()) {
